requirements.txt
uvicorn
numpy<2
piper_tts
sounddevice
faster-whisper
torch==2.7.1
fastapi